added intensity setting for amount of words removed from each quote

+++ script/quoteHandle.js
@@ -1,24 +1,13 @@
 // Check if LocalStorage has any quotes
-// function quoteHandle() {
-//     var data = localStorage.getItem('quotes') ?? null;
-//     console.log(data)
-//     if (data === null) {
-//         // No Quotes
-//         return;
-//     }
-
-//     // Quotes Exist
-
-// }
-
-// quoteHandle();
 
 // Check if url has quotes
 (function () {
 
     var url = new URL(window.location.href);
     var quotes = url.searchParams.get("quotes");
+    var currentPage = url.pathname.split("/").pop().split(".")[0];
 
+    console.log(quotes)
     function handleResults(score, length) {
         var quoteResults = document.getElementById("quoteResults");
         quoteResults.innerHTML = "Score: " + score + "/" + length;
@@ -57,24 +46,92 @@
     function containsSpecialCharacters(word) {
         var specialCharacters = /([.,\/#!$%\^&\*;:{}=\-_`~()?'"]+)/g;
         return specialCharacters.test(word);
-
     }
 
-    if (quotes === null) {
+    function removeWords(randomInput, index, quoteIndex) {
+        // Initialise array
+        if (inputdata[quoteIndex] === undefined) {
+            inputdata[quoteIndex] = {};
+        }
+
+        var random = randomInput;
+        var attempts = 0;
+
+        while ((elementQuote[random] === "" ||
+            elementQuote[random] === " " ||
+            elementQuote[random] === "<br>" ||
+            elementQuote[random] === "##BR##" ||
+            containsSpecialCharacters(elementQuote[random]) ||
+            elementQuote[random].includes("<input")) && attempts < elementQuote.length) {
+            random = Math.floor(Math.random() * elementQuote.length);
+            attempts++;
+        }
+
+        if (attempts === elementQuote.length) {
+            console.log(elementQuote)
+            console.error('Could not find a valid word to remove');
+            return;
+        }
+        var removedWord = elementQuote[random];
+        var inputLength = elementQuote[random].length;
+        console.log("aaaa: " + elementQuote[random])
+
+        if (inputLength < 5) {
+            inputLength = 5;
+        }
+
+        console.log("---- elementQuote2 ----")
+        console.log(elementQuote)
+
+        if (elementQuote[random].endsWith("<br>"))
+            elementQuote[random] = "<input style='width:" + (inputLength) + "em;' class='quoteCheck' placeholder='...' id=quoteCheck-" + quoteIndex + "-" + index + " ><br>";
+        else
+            elementQuote[random] = "<input style='width:" + (inputLength) + "em;' class='quoteCheck' placeholder='...' id=quoteCheck-" + quoteIndex + "-" + index + " >";
+
+        console.log(removedWord)
+
+        // console.log(elementQuote)
+
+        console.log("Includes input?? " + removedWord.includes("<input"))
+
+        console.log(quoteIndex, index)
+        inputdata[quoteIndex][index] = removedWord.replaceAll("<br>", "").replaceAll("##BR##", "");
+        var quote = elementQuote.join(" ").replace(/##BR##/g, "<br>");
+        template.innerHTML = (quoteIndex + 1) + ". " + quote;
+        quoteList.appendChild(template);
+    }
+
+
+
+    if (quotes === null && currentPage === "index") {
+        var intensity = document.getElementById("intensityRange")
+        var intensityOutput = document.getElementById("intensityValue")
+        intensityOutput.innerHTML = intensity.value;
+
+        intensity.oninput = function () {
+            intensityOutput.innerHTML = this.value;
+
+            console.log(this.value);
+
+        }
         // Check if LocalStorage has any quotes
         var data = localStorage.getItem('quotes') ?? null;
-        if (data !== null) {
+
+        if (data !== null && data.length > 0) {
             // Quotes Exist show them
             // First empty the quoteList
             var quoteList = document.getElementById("quoteList");
             quoteList.innerHTML = "";
             var quotes = JSON.parse(decodeURIComponent(atob(data)));
-            var quoteList = document.getElementById("quoteList");
 
             for (var i = 0; i < quotes.length; i++) {
                 addQuoteElement(quoteList, quotes[i]);
             }
         }
+
+        var intensitySetting = localStorage.getItem("intensity") ?? 50;
+        intensity.value = intensitySetting;
+        intensityOutput.innerHTML = intensitySetting;
 
         document.getElementById("addQuote").addEventListener("click", function () {
             var quoteList = document.getElementById("quoteList");
@@ -89,13 +146,43 @@
                 quoteArray.push(quotes[i].value);
             }
 
+            // If quotes are empty
+            if (quoteArray.join("").trim().replaceAll(" ", "") === "") {
+                alert("Please make sure you submit some quotes.");
+                return;
+            }
+
             localStorage.setItem("quotes", btoa(encodeURIComponent(JSON.stringify(quoteArray))));
+            localStorage.setItem("intensity", intensity.value)
+
             window.location.href = "quoteTest.html?quotes=true";
         });
-    } else {
+
+        var optionText = document.getElementsByClassName("optionText")[0]
+        var optionField = document.getElementsByClassName("optionsField")[0]
+
+        optionText.addEventListener("click", function (element) {
+            console.log("Click")
+            if (element.target.innerHTML.includes("▼")) {
+                element.target.innerHTML = "⚙️ Options ▲";
+                optionField.style.display = "block";
+                optionText.classList.remove("optionsClosed");
+                optionText.classList.add("optionsOpened");
+
+            }
+            else {
+                element.target.innerHTML = "⚙️ Options ▼";
+                optionField.style.display = "none";
+                optionText.classList.remove("optionsOpened");
+                optionText.classList.add("optionsClosed");
+
+            }
+        });
+    } else if (currentPage === "quoteTest" && quotes !== null) {
+        // Is the quoteTest page
         var data = decodeURIComponent(atob(localStorage.getItem("quotes")));
+        var intensity = localStorage.getItem("intensity")
         var inputdata = {};
-        var results = {};
 
         if (data === null) {
             // No Quotes
@@ -105,63 +192,102 @@
 
         // Setup Listeners
         var submitButton = document.getElementsByClassName("submitButton")[0];
+
+        // Submit Button Listener
         submitButton.addEventListener("click", function () {
             console.log("Click")
-            results = {};
+            var results = {};
 
             // Check if all inputs are filled
             var quotes = document.getElementsByClassName("quoteCheck");
-
             for (var i = 0; i < quotes.length; i++) {
-                var quote = document.getElementById("quoteCheck-" + i);
-                if (quote.value === "") {
+                if (quotes[i].value === "") {
                     alert("Please fill in all the words to see your score.");
                     return;
                 }
             }
 
-            for (var i = 0; i < quotes.length; i++) {
-                var quote = document.getElementById("quoteCheck-" + i);
-                console.log(inputdata)
-                console.log("Quote: " + quote.value + " Input: " + inputdata[i]);
-                quote.value = quote.value.replaceAll(" ", "");
-                if (quote.value.toLowerCase() !== inputdata[i].toLowerCase()) {
-                    results[i] = false;
-                    quote.style.backgroundColor = "red";
-                    continue;
+            // Check if all inputs are correct
+            for (var j = 0; j < quotes.length; j++) {
+                var quote = quotes[j];
+                console.log(quote.id.split("-"))
+                var index = quote.id.split("-")[2];
+                var i = quote.id.split("-")[1];
+
+                if (results[i] === undefined) {
+                    results[i] = {};
                 }
-                quote.style.backgroundColor = "green";
-                results[i] = true;
-                continue;
+
+                console.log(inputdata)
+                console.log("Quote: " + quote.value + " Input: " + inputdata[i][index]);
+                quote.value = quote.value.replaceAll(" ", "");
+
+                if (quote.value.toLowerCase() !== inputdata[i][index].toLowerCase()) {
+                    results[i][index] = false;
+                    quote.style.backgroundColor = "red";
+                    // Add tooltype to show the correct word
+                    quote.setAttribute("title", 'Correct Word: "' + inputdata[i][index] + '"');
+
+                }
+                else {
+                    quote.style.backgroundColor = "green";
+                    results[i][index] = true;
+                }
             }
+
+            // Calculate Score and show it
             var score = 0;
-            for (var i = 0; i < quotes.length; i++) {
-                if (results[i] === true) {
+            console.log("Lenght: " + Object.keys(results).length)
+            // for (var j = 0; j < quotes.length; j++) {
+            //     var quote = quotes[j];
+            //     var index = quote.id.split("-")[2];
+            //     var i = quote.id.split("-")[1];
+            //     if (results[i][index] === true) {
+            //         score++;
+            //     }
+            // }
+
+            for (var key in results) {
+                var allTrue = true;
+                for (var key2 in results[key]) {
+                    if (results[key][key2] === false) {
+                        allTrue = false;
+                    }
+                }
+                if (allTrue === true) {
                     score++;
                 }
             }
-            handleResults(score, quotes.length);
+
+            handleResults(score, Object.keys(results).length + 1);
 
         });
 
+        // Back Button Listener
         document.getElementsByClassName("backButton")[0].addEventListener("click", function () {
             window.location.href = "index.html";
         });
 
+        // Quote Setup Sequence
         var quotes = JSON.parse(data);
         var quoteList = document.getElementById("quoteList");
+
         for (var i = 0; i < quotes.length; i++) {
+
             // Quote text show
             var template = document.createElement("div");
             template.setAttribute("class", "quoteShow");
+
             // Check if it has a enter character in it
             if (quotes[i].includes("\n")) {
                 quotes[i] = quotes[i].replaceAll("\n", "<br>");
             }
+
             var elementQuote = quotes[i].split(" ");
             console.log("---- elementQuote ----")
             console.log(elementQuote)
-            // If word includes <br> split 
+
+            // If word includes <br> split it and use ##BR## to replace it later
             elementQuote.forEach((word, index) => {
                 if (word.includes("<br>")) {
                     console.log(index)
@@ -176,45 +302,48 @@
             elementQuote = elementQuote.map(word => word.replace("<br>", "##BR##"));
 
             // Split special characters from words and add them to the array but dont seperate each individual character
-            elementQuote.forEach((word, index) => {
+            let newElementQuote = [];
+
+            elementQuote.forEach((word) => {
                 if (containsSpecialCharacters(word)) {
-                    var splitWord = word.split(/([.,\/#!$%\^&\*;:{}=\-_`~()?'"]+)/g);
-                    console.log("-> " + splitWord)
-                    elementQuote.splice(index, 1, ...splitWord.filter(Boolean));
+                    let splitWord = word.split(/([.,\/#!$%\^&\*;:{}=\-_`~()?'"]+)/g);
+                    newElementQuote.push(...splitWord.filter(Boolean));
+                } else {
+                    newElementQuote.push(word);
                 }
             });
+            
+            elementQuote = newElementQuote;
 
             console.log("---- elementQuote1 ----")
             console.log(elementQuote)
-            // Randomly remove 1 word from the quote
-            var random = Math.floor(Math.random() * elementQuote.length);
 
 
-            while (elementQuote[random] === "" || elementQuote[random] === " " || elementQuote[random] === "<br>" || elementQuote[random] === "##BR##" || containsSpecialCharacters(elementQuote[random])) {
-                random = Math.floor(Math.random() * elementQuote.length);
+            // Randomly remove certain amount of words from the quote
+            if (intensity < 10) {
+                var wordsRemoved = Math.ceil(elementQuote.length * (intensity / 100));
+            }
+            else {
+                var wordsRemoved = Math.floor(elementQuote.length * (intensity / 100));
             }
 
-            var removedWord = elementQuote[random];
-            var inputLength = elementQuote[random].length;
 
-            if (inputLength < 5) {
-                inputLength = 5;
+            console.log("Removing: " + wordsRemoved + " words" + " from " + elementQuote.length + " words")
+
+
+            for (var j = 0; j < wordsRemoved; j++) {
+                var random = Math.floor(Math.random() * elementQuote.length);
+                removeWords(random, j, i);
             }
-            console.log("---- elementQuote2 ----")
-            console.log(elementQuote)
-            if (elementQuote[random].endsWith("<br>"))
-                elementQuote[random] = "<input style='width:" + (inputLength) + "em;' class='quoteCheck' placeholder='...' id=quoteCheck-" + i + " ><br>";
-            else
-                elementQuote[random] = "<input style='width:" + (inputLength) + "em;' class='quoteCheck' placeholder='...' id=quoteCheck-" + i + " >";
-            console.log(removedWord)
-            // console.log(elementQuote)
-            inputdata[i] = removedWord.replaceAll("<br>", "").replaceAll("##BR##", "");
-            var quote = elementQuote.join(" ").replace(/##BR##/g, "<br>");
-            template.innerHTML = (i + 1) + ". " + quote;
-            quoteList.appendChild(template);
+
+
         }
         console.log("---- Quotes ----")
         console.log(quotes)
+    }
+    else {
+        var box = document.getElementsByClassName("border-box")[0]
+        box.innerHTML = "<a href='./index.html' class='backButton'>&#8592; Go Back</a> <h1 style='padding-top:20px'>Something Went Wrong...</h1><p>Please go back and make sure you submit your quotes.</p>"
     }
 
 })();
